Added support for Keyword reading
Keyword support has been adding in. Upon creating the parser, A file containing the keywords and their ID is read into a vector, When the lexer encounters an Alpha char, it passes control over to the keywordCheck function which reads all characters in until a delimiter at which point it checks the resulting string against the keyword list, registering it as a keyword token if a match is found, or as a identifier token if no match is found.