Fixed IDs for tokens, floats and ints
The operator IDs should be fine. Also noticed the floats and ints weren't getting the proper ID because of the way I laid out the state table, so I added a new function for that. Also added to the scanFile function an AddToken call for end of file after the while loop. There is one strange one bug if scanning "Ex2.txt" where it finds the character "\r" at the end of file. It doesn't know what to do with it so it assigns it ID 99 (error). None of the other text files had this character at the end.