Updated to new version
Updated to run through a menu which is stored in Compiler.h. All functionality is now run through the Compiler class and other classes have been modified to work with this. Each class also now has an "error code" int variable which is set if an error is encountered, upon returning to the compiler class, this error can be referenced and output.